storing percentage position for boulders on floormap
- changed positioning of markers (no more complex JS calculation, but offset of input elements inside li elements)

+++ grails-app/domain/bcomp/gym/Boulder.groovy
@@ -15,7 +15,7 @@ class Boulder {
 
     String grade
 
-    public void onFloorPlan(FloorPlan floorPlan, int x, int y) {
+    public void onFloorPlan(FloorPlan floorPlan, double x, double y) {
         location = new OnFloorPlan(floorPlan: floorPlan, x: x, y: y)
         location.boulder = this
     }
